Small fixes.
	* mln/accu/math/inf.hh,
	* mln/accu/math/sup.hh,
	* mln/accu/shape/height.hh,
	* mln/accu/shape/volume.hh,
	* mln/accu/stat/median_alt.hh,
	* mln/value/stack.hh: fix invalid namespace closing.

	* mln/util/internal/graph_iter.hh,
	* mln/fun/v2v/rgb_to_hsl.hh: fix ambiguities with operators.

	* mln/value/internal/gray_f.hh: repair header.

	* tests/registration/Makefile.am: disable registration test.

git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@4150 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/fun/v2v/rgb_to_hsl.hh
@@ -1,4 +1,5 @@
-// Copyright (C) 2008, 2009 EPITA Research and Development Laboratory (LRDE)
+// Copyright (C) 2008, 2009 EPITA Research and Development Laboratory
+// (LRDE)
 //
 // This file is part of Olena.
 //
@@ -102,7 +103,7 @@ namespace mln
 	    else
 	      hsl.hue() = (60. * (rgb.red() - rgb.green()) / (rmax - rmin)) + 240;
 
-	hsl.lum() = ((double) rmax + rmin) / 2;
+	hsl.lum() = ((double) rmax + (double) rmin) / 2;
 
 	// We want min and max between 0 and 1
 	rmax -= mln_min(typename T_rgb::red_t);
